requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1
django-allauth==65.1.0
django-ckeditor==6.7.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
whitenoise==6.8.2
requests
cryptography